requirements.txt
openai
langchain
faiss-cpu
python-dotenv
PyMuPDF
pytesseract
Pillow
python-pptx
fpdf2
tiktoken
langchain-community 
langchain-openai
docx2pdf
python-docx
markdown2
beautifulsoup4
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-docx
docx2pdf
markdown2
beautifulsoup4
pytesseract
Pillow
python-pptx
PyMuPDF
python-dotenv
python-docx